+++ firebase/functions/package.json
@@ -13,7 +13,7 @@
     "node": "8"
   },
   "dependencies": {
-    "apollo-server-express": "^2.5.0",
+    "apollo-server-cloud-functions": "^2.5.0",
     "body-parser": "^1.19.0",
     "cors": "^2.8.5",
     "express": "^4.17.0",
